Simplify EIP-6110 Deposit requests
